+++ src/components/RatingComponent.js
@@ -8,6 +8,7 @@ export default function RatingComponent(props) {
   const getRatingList = props.getRatingList;
 
   const [ratingModifyOpen, setRatingModifyOpen] = useState(false);
+  const [isLoading, setLoading] = useState(false);
 
   const showRatingModify = () => {
     setRatingModifyOpen(true);
@@ -26,22 +27,42 @@ export default function RatingComponent(props) {
       },
     };
 
+    const yesOrNo = window.confirm('등급을 삭제하시겠습니까?');
+    if (yesOrNo === false) {
+      return;
+    }
+
     try {
-      console.log('staffToken', token);
+      setLoading(true);
+
       const response = await serverapi.delete(api, options);
       console.log('response', response.data);
 
+      alert('삭제되었습니다');
       getRatingList();
     } catch (error) {
       console.log(error);
       alert(error.response.data.message);
+    } finally {
+      setLoading(false);
     }
   };
 
   return (
-    <tr key={rating.ratingId}>
+    <tr key={rating.code}>
+      <td>{rating.code}</td>
       <td>{rating.name}</td>
-      <td>{rating.minAge}</td>
+      <td>
+        <button class='btn btn-danger' onClick={deleteRating(rating.code)}>
+          {isLoading ? (
+            <div className='spinner-border' role='status'>
+              <span className='sr-only' />
+            </div>
+          ) : (
+            <span>삭제</span>
+          )}
+        </button>
+      </td>
       <td>
         <button class='btn btn-warning' onClick={showRatingModify}>
           수정
@@ -58,14 +79,6 @@ export default function RatingComponent(props) {
             rating={rating}
           />
         </Modal>
-      </td>
-      <td>
-        <button
-          class='btn btn-danger'
-          onClick={() => deleteRating(rating.ratingId)}
-        >
-          삭제
-        </button>
       </td>
     </tr>
   );
